Mednovice how to merge? naa nani katong ma store na ang manager and date? mehehe

+++ app/Http/Controllers/UsersController.php
@@ -5,8 +5,17 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\User;
 
+use App\Permissions;
+use Log;
+
 class UsersController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware('permissions', ['except' => ['store', 'update', 'destroy']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -81,7 +90,17 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
-        User::destroy($id);
-        return redirect()->back();
+        $data = Permissions::hasDeleteUserPermission($id);
+        
+        if(trim($data) === '')
+        {
+            User::destroy($id);
+            return redirect()->back();
+        }
+        else
+        {
+            Log::error('ACCESS DENIED. User tries to delete data of UserType=' . strtoupper($data) . ' but is not included in the current user\'s list of permissions.');
+            abort(503);
+        }
     }
 }
